Add option to keep previous session as a backup
When enabled, this helps to prevent loosing the session in case of an unexpected OS/app crash. I have many pre-configured tabs, that I already lost many times and I always have to set them up again, which is cumbersome.

I added it as an option to not modify the current (intentional?) behavior. Of course we could also just simply remove the line `GLib.unlink(this.session_file_path);`.