improve session file handling and session restore error handling
Do not delete the session file after successfully restoring it. This has the benefit that we do not loose the session in case of a OS/app crash.

Delete the session file if all tabs are closed (windows destroyed).

Delete the session file in case something goes wrong during restoring the session.