app: don't remove session file after loading it (#1198)
Do not delete the session file after successfully restoring it. This has the benefit that we do not loose the session in case of a OS/app crash.

Delete the session file if all tabs are closed.

Delete the session file in case something goes wrong during restoring the session.

Co-authored-by: Aleksandr Mezin <mezin.alexander@gmail.com>